More rigorous approach
In this version, I've created a `const` for all
string literals used.  In this way, the string
literal appears only in one place.  This has
three big advantages:

  * If you mistype something, the compiler will
    warn you.
  * You can change the values of the string
    literals easily in one place.
  * You can rename the variables themselves
    using VSCode's rename functionality.

I've also replaced all classes with pure
interfaces.  The main reason to do this is
to eliminate the possibility that an `instanceof`
operator might be used and to force and
discrimination to be done using the `type`
property.

I've also created functions to easily create
objects that are instances of these interfaces.
Technically, I really don't typically need the
interface definitions because they can be
inferred directly in the definition of the creator
functions.  But there are some times where I
might want to specifically refer to the type, so
I've included them all for completeness.
Furthermore, it is nice to have a formal definition
of the type for purely documentation purposes.